requirements.txt
requests>=2.31.0
docker>=6.1.1
openai>=1.3.0
python-dotenv>=1.0.0
rich>=13.7.0 